refactor(agents): rename skills resource module to skill_resource
