+++ src/phoenix/server/agents/capabilities/skills/__init__.py
@@ -1,6 +1,6 @@
 from phoenix.server.agents.capabilities.skills.capability import SkillsCapability
 from phoenix.server.agents.capabilities.skills.parsing import parse_skill_md
-from phoenix.server.agents.capabilities.skills.resource import SkillResource
+from phoenix.server.agents.capabilities.skills.skill_resource import SkillResource
 from phoenix.server.agents.capabilities.skills.skill import Skill
 from phoenix.server.agents.capabilities.skills.toolset import SkillsToolset
