+++ src/phoenix/server/agents/capabilities/skills/skill_resource.py
